requirements.txt
streamlit~=1.51.0
pandas~=2.3.3
python-dotenv~=1.1.0
langchain==0.2.14
langchain-core==0.2.39
langchain-community==0.2.12
langchain_groq==0.1.9